hooksに切り分けて責務を分割
